Adapt 'Delete' modal to user groups
The 'Delete' modal functionality
must be adapted to work with the
User groups data.
